clloader classes only control train data or test data + move clloader to abstract class et specifilized the others (wip)
